moved ScaScanService to AnalysisService
